Added logic to prevent a disk from being deleted when it's in a valid status.

+++ http/src/main/scala/org/broadinstitute/dsde/workbench/leonardo/monitor/LeoPubsubMessageSubscriber.scala
@@ -257,6 +257,7 @@ class LeoPubsubMessageSubscriber[F[_]](
   ): F[Unit] =
     for {
       ctx <- ev.ask
+      _ <- logger.info(ctx.loggingCtx)("++++ handleDeleteRuntimeMessage")
       existingOperationFuture <- operationFutureCache.get(msg.runtimeId)
       _ <- existingOperationFuture.traverse(opFuture => F.delay(opFuture.cancel(true)))
       runtimeOpt <- clusterQuery.getClusterById(msg.runtimeId).transaction
@@ -302,6 +303,7 @@ class LeoPubsubMessageSubscriber[F[_]](
           now <- nowInstant
           diskOpt <- persistentDiskQuery.getPersistentDiskRecord(id).transaction
           disk <- F.fromEither(diskOpt.toRight(new RuntimeException(s"disk not found for ${id}")))
+          _ <- logger.info(ctx.loggingCtx)("++++ LeoPubsub delete disk")
           deleteDiskOp <- googleDiskService.deleteDisk(googleProject, disk.zone, disk.name)
           _ <- deleteDiskOp.traverse(x => F.blocking(x.get()))
           _ <- persistentDiskQuery.delete(id, now).transaction.void >> authProvider
@@ -766,6 +768,7 @@ class LeoPubsubMessageSubscriber[F[_]](
         LeoLenses.cloudContextToGoogleProject.get(disk.cloudContext),
         new RuntimeException("non google project cloud context is not supported yet")
       )
+      _ <- logger.info(ctx.loggingCtx)("++++ Direct delete disk")
       opFutureOpt <- googleDiskService.deleteDisk(googleProject, disk.zone, disk.name)
       _ <- opFutureOpt match {
         case None => F.unit
@@ -1336,6 +1339,7 @@ class LeoPubsubMessageSubscriber[F[_]](
   )(implicit ev: Ask[F, AppContext]): F[Unit] =
     for {
       ctx <- ev.ask
+      _ = print("++++ Barpes")
       _ <- metrics.incrementCounter(s"createRuntimeError", 1)
       _ <- logger.error(ctx.loggingCtx, e)(s"Failed to create runtime ${runtimeId}")
       // want to detach persistent disk for runtime
@@ -1343,7 +1347,6 @@ class LeoPubsubMessageSubscriber[F[_]](
         case CloudService.GCE =>
           for {
             runtimeOpt <- clusterQuery.getClusterById(runtimeId).transaction
-
             _ <- runtimeOpt.traverse_ { runtime =>
               // If the disk is in Creating status, then it means it hasn't been used previously. Hence delete the disk
               // if the runtime fails to create.
@@ -1365,11 +1368,17 @@ class LeoPubsubMessageSubscriber[F[_]](
                       persistentDiskOpt <- rc.persistentDiskId.flatTraverse(did =>
                         persistentDiskQuery.getPersistentDiskRecord(did).transaction
                       )
-                      _ <- persistentDiskOpt.traverse_(d =>
-                        googleDiskService.deleteDisk(googleProject, rc.zone, d.name) >> persistentDiskQuery
-                          .updateStatus(d.id, DiskStatus.Deleted, now)
-                          .transaction
-                      )
+                      _ = persistentDiskOpt match {
+                        case Some(value) =>
+                          if (value.status == DiskStatus.Creating || value.status == DiskStatus.Failed) {
+                            persistentDiskOpt.traverse_(d =>
+                              googleDiskService.deleteDisk(googleProject, rc.zone, d.name) >> persistentDiskQuery
+                                .updateStatus(d.id, DiskStatus.Deleted, now)
+                                .transaction
+                            )
+                          }
+                        case None => F.unit
+                      }
                     } yield ()
                   }
                 } yield ()
